requirements.txt
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
dominate==2.3.5
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-DotEnv==0.1.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.2
visitor==0.1.3
Werkzeug==0.15.2
